finished thumbs movement + bar coloring

+++ util/src/main/java/com/jueggs/customview/util/Utils.kt
@@ -35,4 +35,4 @@ fun Context.getColorCompat(@ColorRes resId: Int) = ContextCompat.getColor(this, 
 
 fun isLollipopOrAbove(): Boolean = Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP
 
-fun rangeCrop(bottomLimit: Int, topLimit: Int, value: Int) = Math.min(Math.max(bottomLimit, value), topLimit)
+fun rangeCrop(bottomLimit: Float, topLimit: Float, value: Float) = Math.min(Math.max(bottomLimit, value), topLimit)
